requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
alabaster
async-timeout
asynctest
attrs
awscli
babel
bcrypt
boto
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
contourpy
corider
cryptography
cycler
dataclasses
decorator
Deprecated
docutils
et-xmlfile
fabric
filelock
fonttools
frozenlist
fsspec
future
google-auth
google-auth-oauthlib
GPUtil
grpcio
h11
humanfriendly
idna
idna-ssl
imagesize
importlib_metadata
importlib_resources
invoke
Jinja2
jmespath
joblib
kiwisolver
Markdown
MarkupSafe
matplotlib
MouseInfo
mpmath
multidict
mysql-connector-python
natsort
networkx
numpy
oauthlib
openpyxl
outcome
packaging
pandas
paramiko
pathlib2
pillow
prettytable
protobuf
psutil
ptflops
py-cpuinfo
pyasn1
pyasn1_modules
PyAutoGUI
pycparser
pyDeprecate
PyGetWindow
Pygments
PyMsgBox
pymysql
PyNaCl
pyparsing
pyperclip
Pypubsub
pyreadline
pyreadline3
PyRect
PyScreeze
PySocks
python-dateutil
python-dotenv
pytorch-benchmark
pytorch-lightning
pytweening
pytz
pywin32
PyYAML
requests
requests-oauthlib
ride
robotframework
robotframework-metrics
robotframework-pabot
robotframework-pythonlibcore
robotframework-requests
robotframework-ride
robotframework-seleniumlibrary
robotframework-stacktrace
rsa
s3transfer
scipy
seaborn
selenium
six
sniffio
snowballstemmer
sortedcontainers
Sphinx
sphinx-rtd-theme
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jquery
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
supers
sympy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
torch
torchmetrics
tqdm
trio
trio-websocket
typing_extensions
tzdata
urllib3
wcwidth
websocket-client
Werkzeug
wrapt
wsproto
wxPython
xmltodict
yarl
zipp